No sound after changing to interpolation and ms, commit before debugging

+++ Blok2d/opdrachten/00_combReverb/combFilter.h
@@ -4,31 +4,28 @@
 
 #pragma once
 
-#include <iostream>
-#include <vector>
 #include "circularBuffer.h"
 
 class CombFilter{
 public:
   CombFilter() = default;
-  CombFilter(unsigned int D, float g);
+  CombFilter(float D, float g);
   ~CombFilter();
 
   double process(double input);
 
-  void setD(unsigned int D);
-  unsigned int getD();
+  void setD(float D);
+  float getD();
   //snakecase because g should be lowercase
   void set_g(float g);
   float get_g();
 
 private:
-
   CircularBuffer x;
   CircularBuffer y;
 
   //D is the delay in the combfilter, also the bufferSize (Pirkle p464)
-  unsigned int D;
+  float D;
   //g is the feedback value (Pirkle p464)
   float g;
